open_food_facts/requirements.txt
django==3.0
djangorestframework==3.12.0
djongo==1.3.4
dnspython==2.1.0
requests==2.25.1
psutil==5.8.0